requirements.txt
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
django-filter==23.2
psycopg2-binary==2.9.6